feat: create ResultPage using expo router and implement useQuery to get fipe values
